pyproject.toml
[tool.poetry]
name = "codeine"
version = "0.1.0"
description = ""
authors = ["SlapDrone <slap@drone.io>"]
readme = "README.md"

[tool.poe.tasks]
track = "mlflow server --backend-store-uri sqlite:///mlflow.db --default-artifact-root ./artifacts --host 0.0.0.0"
track_old = "mlflow ui -h 0.0.0.0 -p 5000"                                
jn = "jupyter-notebook --no-browser --allow-root --ip 0.0.0.0 --port 8888"

[tool.poetry.dependencies]
python = "3.10.8"
gradio = "*"
mdtex2html = "*"
pypinyin = "*"
tiktoken = "*"
socksio = "*"
tqdm = "^4.65.0"
colorama = "*"
duckduckgo_search = "*"
Pygments = "*"
llama_index = "*"
langchain = "*"
markdown = "*"
markdown2 = "*"
SentencePiece = "*"
pandas = "*"
addict = "*"
poethepoet = "^0.20.0"
seaborn = "^0.12.2"
python-dotenv = "^1.0.0"
aioitertools = "^0.11.0"

[tool.poetry.dev-dependencies]
jupyter="*"
seaborn="*"
matplotlib="*"
jupyterthemes="*"
pytest="*"
pylint="*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
